backend/requirements-experiment.txt
requests
python-dotenv
netCDF4
numpy
